move prune rgb util to state package
