+++ app/packages/app/src/useSetters/onSetColorScheme.ts
@@ -5,11 +5,10 @@ import {
   setColorScheme,
   setColorSchemeMutation,
 } from "@fiftyone/relay";
-import { ensureColorScheme } from "@fiftyone/state";
+import { ensureColorScheme, removeRgbProperty } from "@fiftyone/state";
 import { DefaultValue } from "recoil";
 import { commitMutation } from "relay-runtime";
 import { RegisteredSetter } from "./registerSetter";
-import { cloneDeep } from "lodash";
 
 const onSetColorScheme: RegisteredSetter =
   ({ environment, setter, subscription }) =>
@@ -41,26 +40,3 @@ const onSetColorScheme: RegisteredSetter =
   };
 
 export default onSetColorScheme;
-
-export function removeRgbProperty(input) {
-  // Clone the input to avoid mutating the original object
-  const clonedInput = cloneDeep(input);
-
-  // Process the 'colorscales' array
-  if (clonedInput.colorscales && Array.isArray(clonedInput.colorscales)) {
-    clonedInput.colorscales = clonedInput.colorscales.map(
-      ({ rgb, ...rest }) => rest
-    );
-  }
-
-  // Process the 'defaultColorscale' object
-  if (
-    clonedInput.defaultColorscale &&
-    typeof clonedInput.defaultColorscale === "object"
-  ) {
-    const { rgb, ...rest } = clonedInput.defaultColorscale;
-    clonedInput.defaultColorscale = rest;
-  }
-
-  return clonedInput;
-}
